requirements.txt
fastapi==0.128.0

gunicorn==23.0.0
uvicorn==0.40.0

python-multipart==0.0.21
python-dotenv==1.2.1
PyYAML==6.0.3

google-cloud-storage==2.19.0
google-cloud-firestore==2.22.0
